+ Update some recipes to work with latest versions / bookworm. + New "del_last_zss.sh" script to remove last state snapshot. + Fix setup & first_boot scripts.

+++ scripts/recipes/install_gula.sh
@@ -2,11 +2,15 @@
 
 # install_gula.sh
 cd $ZYNTHIAN_PLUGINS_SRC_DIR
-rm -rf gula-plugins
+
+if [ -d "gula-plugins" ]; then
+	rm -rf gula-plugins
+fi
+
 git clone https://github.com/steveb/gula-plugins.git
 cd gula-plugins
-
 make all
 cp -r lv2/* $ZYNTHIAN_PLUGINS_DIR/lv2/
-
 cd ..
+
+rm -rf gula-plugins
